requirements.txt
ccxt>=4.1.13
pandas>=2.0.0
numpy>=1.24.0
tensorflow-cpu>=2.13.0
scikit-learn>=1.3.0
ta>=0.10.2
python-dateutil>=2.8.2
openpyxl>=3.1.2  # Required for Excel operations
python-dotenv>=1.0.0  # For environment variable management 